Camera 클래스 추가 및 framework에서 렌더링하는 모든 과정을 작성
Camera 클래스는 GameObject를 상속받고 따로 카메라와 관련된 변수들을 가진다. 각 변수에 대한 것은 주석 참고,
Camera를 GameObject에 Child로 쉽게 종속시키기 GameObject를 상속받았다. 주의 할 점으로 Camera는 충돌처리와 렌더링을 하지 않도록 오버라이딩 해야한다.

GameFramework의 FrameAdvance() 의 내용 추가:
1.렌더링을 하기 위해 리소스 배리어를 설정하고
2.렌더 타겟 버퍼와 깊이-스텐실 버퍼 의 CPU주소를 계산하고
3.기본 배경 색상으로 초기화하고
4.그래픽 루트 시그니처를 Set하고
5.씬을 렌더링
6.리소스 배리어를 통해 렌더링(쓰는 과정)이 끝나길 기다린다.
7.CPU와 GPU동기화
8.그려진 후면버퍼를 전면버퍼로 프리젠트
9.다음 후면버퍼로 이동하고 CPU와 GPU동기화
위와 같이 Framework에서 렌더링을 위한 전 과정을 추가하였다.

GameObject에서 wrold좌표계 기준으로 자신의 right, up, look, position 벡터를 얻을 수 있는 get함수를 작성 하였다.

Scene의 FrameAdvance() 함수를 추상함수(자식 클래스에서 반드시 정의하도록 하는 선언한 함수)로 바꿈
Scene에서 직접 플레이어와 카메라를 관리'제어 하도록 한다.

stdafx파일 에서 리소스를 생성(할당)하는 함수를 작성:
힙타입이 디폴트일경우 업로드 힙을 생성하고 디폴트힙으로 카피한다.

GameFramework의 변수 이름을 통일성있게 수정

+++ MainProject/MainProject/GameTimer.cpp
@@ -41,7 +41,7 @@ int GameTimer::GetFPS() const
 		cout << "GameTimer::GetFPS() : 일시정지 상태에서 함수를 호출했습니다.\n";
 #endif // DEBUG
 
-	return m_frameTimes.size() + 1;
+	return (int)m_frameTimes.size() + 1;
 }
 bool GameTimer::IsPaused() const
 {
